# web/ui/openapi.json
{
  "components": {
    "parameters": {
...
      "BucketPath": {
        "in": "path",
        "name": "bucket",
        "required": true,
        "schema": {
          "enum": [
            2,
            30,
            60,
            120,
            180,
            300
          ],
          "type": "integer"
        }
      },
...
      "CountPath": {
        "in": "path",
        "name": "count",
        "required": true,
        "schema": {
          "minimum": 1,
          "type": "integer"
        }
      },
...
      "FileNamePath": {
        "in": "path",
        "name": "fileName",
        "required": true,
        "schema": {
          "pattern": "^[a-zA-Z0-9_\\\\-.]+$",
          "type": "string"
        }
      },
...
      "VersionPath": {
        "in": "path",
        "name": "version",
        "required": true,
        "schema": {
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "requestBodies": {
...
      "EchCertForm": {
        "content": {
          "application/x-www-form-urlencoded": {
            "schema": {
              "properties": {
                "sni": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "required": false
      },
      "ImportDatabaseMultipart": {
        "content": {
          "multipart/form-data": {
            "schema": {
              "properties": {
                "db": {
                  "description": ".db, .sqlite, or .sqlite3 backup file within the server size limit.",
                  "format": "binary",
                  "type": "string"
                }
              },
              "required": [
                "db"
              ],
              "type": "object"
            }
          }
        },
        "required": true
      },
...
      "LogsFilterForm": {
        "content": {
          "application/x-www-form-urlencoded": {
            "schema": {
              "properties": {
                "level": {
                  "type": "string"
                },
                "syslog": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "required": false
      },
...
      "XrayLogsFilterForm": {
        "content": {
          "application/x-www-form-urlencoded": {
            "schema": {
              "properties": {
                "filter": {
                  "type": "string"
                },
                "showBlocked": {
                  "type": "string"
                },
                "showDirect": {
                  "type": "string"
                },
                "showProxy": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        },
        "required": false
      },
...
    },
    "responses": {
      "AuthNotFound": {
        "description": "Unauthenticated /panel/api requests intentionally return 404 to hide API existence."
      },
      "BinaryFile": {
        "content": {
          "application/octet-stream": {
            "schema": {
              "format": "binary",
              "type": "string"
            }
          }
        },
        "description": "Authenticated binary file download."
      },
      "ForbiddenCSRF": {
        "content": {
          "application/json": {
            "examples": {
              "csrfFailed": {
                "value": {
                  "msg": "CSRF validation failed",
                  "obj": null,
                  "success": false
                }
              }
            },
            "schema": {
              "$ref": "#/components/schemas/LegacyResponse"
            }
          }
        },
        "description": "CSRF validation failed for a state-changing request."
      },
      "LegacyMsg": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LegacyResponse"
            }
          }
        },
        "description": "Legacy JSON response envelope used by web/entity.Msg."
      }
    },
    "schemas": {
...
      "LegacyResponse": {
        "properties": {
          "msg": {
            "type": "string"
          },
          "obj": {
            "description": "Endpoint-specific payload. Existing clients unwrap this field."
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "msg",
          "obj"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "description": "HttpOnly SameSite=Lax panel session cookie issued by POST /login.",
        "in": "cookie",
        "name": "SuperXray",
        "type": "apiKey"
      },
      "csrfToken": {
        "description": "Session-bound CSRF token injected into the Vue runtime config; X-XSRF-Token is accepted as a compatibility alias.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Machine-readable contract for the Vue panel API implemented by the local Go/Gin backend.\nRuntime deployments may prepend the configured webBasePath before these paths.\nThe legacy response envelope is web/entity.Msg: success, msg, obj.\n",
    "title": "SuperXray Panel API",
    "version": "3.3.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/panel/api/server/cpuHistory/{bucket}": {
      "get": {
        "operationId": "getCpuHistory",
        "parameters": [
          {
            "$ref": "#/components/parameters/BucketPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Get aggregated CPU history for an allowed bucket size.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/getConfigJson": {
      "get": {
        "operationId": "getConfigJson",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Get the generated Xray configuration JSON.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/getDb": {
      "get": {
        "operationId": "downloadDatabase",
        "responses": {
          "200": {
            "$ref": "#/components/responses/BinaryFile"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Download the authenticated panel database backup.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/getNewEchCert": {
      "post": {
        "operationId": "getNewEchCert",
        "requestBody": {
          "$ref": "#/components/requestBodies/EchCertForm"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenCSRF"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "summary": "Generate an ECH certificate helper value for the provided SNI.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/getNewUUID": {
      "get": {
        "operationId": "getNewUUID",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Generate a new UUID helper value.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/getNewVlessEnc": {
      "get": {
        "operationId": "getNewVlessEnc",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Generate a new VLESS encryption helper value.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/getNewX25519Cert": {
      "get": {
        "operationId": "getNewX25519Cert",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Generate a new X25519 helper value.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/getNewmldsa65": {
      "get": {
        "operationId": "getNewMldsa65",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Generate a new ML-DSA-65 helper value.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/getNewmlkem768": {
      "get": {
        "operationId": "getNewMlkem768",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Generate a new ML-KEM-768 helper value.",
        "tags": [
          "Server"
        ]
      }
    },
...
    "/panel/api/server/importDB": {
      "post": {
        "operationId": "importDatabase",
        "requestBody": {
          "$ref": "#/components/requestBodies/ImportDatabaseMultipart"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenCSRF"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "summary": "Import a database backup file after size/name/extension validation.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/installXray/{version}": {
      "post": {
        "operationId": "installXray",
        "parameters": [
          {
            "$ref": "#/components/parameters/VersionPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenCSRF"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "summary": "Install or update Xray to a selected version.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/logs/{count}": {
      "post": {
        "operationId": "getPanelLogs",
        "parameters": [
          {
            "$ref": "#/components/parameters/CountPath"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/LogsFilterForm"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenCSRF"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "summary": "Read panel logs with optional form filters.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/restartXrayService": {
      "post": {
        "operationId": "restartXrayService",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenCSRF"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "summary": "Restart the legacy Xray service path. This is not routed through CoreManager.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/status": {
      "get": {
        "operationId": "getServerStatus",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Get cached panel, host, network, and Xray runtime status.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/stopXrayService": {
      "post": {
        "operationId": "stopXrayService",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenCSRF"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "summary": "Stop the legacy Xray service path. This is not routed through CoreManager.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/updateGeofile": {
      "post": {
        "operationId": "updateAllGeofiles",
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenCSRF"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "summary": "Update default geofiles.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/updateGeofile/{fileName}": {
      "post": {
        "operationId": "updateGeofile",
        "parameters": [
          {
            "$ref": "#/components/parameters/FileNamePath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenCSRF"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "summary": "Update one allowlisted geofile by filename.",
        "tags": [
          "Server"
        ]
      }
    },
    "/panel/api/server/xraylogs/{count}": {
      "post": {
        "operationId": "getXrayLogs",
        "parameters": [
          {
            "$ref": "#/components/parameters/CountPath"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/XrayLogsFilterForm"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/LegacyMsg"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenCSRF"
          },
          "404": {
            "$ref": "#/components/responses/AuthNotFound"
          }
        },
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "summary": "Read Xray logs with optional form filters.",
        "tags": [
          "Server"
        ]
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "servers": [
    {
      "description": "Runtime panel base path. Use / for the source default; deployments may use /xui/ or another configured prefix.",
      "url": "{webBasePath}",
      "variables": {
        "webBasePath": {
          "default": "/"
        }
      }
    }
  ],
...
}